put `add-opens` in acceptance tests to fix tests
Acceptance and docker acceptance tests use the system jruby to run the tests.
Moving to jdk17 requires the version of Java used to have extra `add-opens` when
starting:

--add-opens=java.base/java.security=ALL-UNNAMED

is required to enable Gem::Security to work correctly

and

--add-opens=java.base/sun.nio.ch=ALL-UNNAMED --add-opens=java.base/java.io=ALL-UNNAMED

to workaround https://github.com/jruby/jruby/issues/6726 on the older jruby
on the build nodes.